requirement.txt
pytorch==1.8.2+cpu
numpy==1.21.5